feat/abstract-simulation-refactor: refactor AS out; redo merge with proper merge message
